web-api: refactored WebSocketStompTransport
